Nest the Apk, Cbor, Mathml and Pem families
`ApkConfiguration`, `ApkManifest` and `ApkSigner` become `Apk.*`; `CborError` and
`CborReader` become `Cbor.*`; `MathmlError` and `MathmlParser` become `Mathml.*`;
`PemError` and `PemLabel` become `Pem.*`. Nine names leave the umbrella's exported
surface.

`MathmlReader` stays: its `searchNodes` indexes an `Array[honeycomb.Node]^{}`,
which the enclosing object requires at `^{any}` — the read-only-versus-exclusive
capture failure that also blocks `TarHeader`. It is the second confirmed instance,
and both are `Array` indexing moved into a capture-impure enclosing scope.

`CborError` was exported from breviloquence's *parser* component while `Cbor` is
in core; the nested member is reachable through core's export, so the parser
export file just drops it.
