--- lib/anthology/src/apk/anthology.ApkConfiguration.scala
@@ -1,71 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package anthology
-
-import anticipation.*
-import gossamer.*
-
-object ApkConfiguration:
-  // The conventional Android debug keystore (shared with Android Studio), a PKCS#12 store with a
-  // well-known alias and password. The defaults produce a debuggable, emulator-installable APK.
-  def default: ApkConfiguration =
-    val home = java.lang.System.getProperty("user.home").nn.tt
-
-    ApkConfiguration
-      ( minApi      = 26,
-        targetApi   = 34,
-        packageName = t"dev.soundness.app",
-        versionCode = 1,
-        versionName = t"1.0",
-        label       = t"Soundness App",
-        permissions = Nil,
-        keystore    = t"$home/.android/debug.keystore",
-        storePass   = t"android",
-        alias       = t"androiddebugkey",
-        keyPass     = t"android" )
-
-// What an Android application package needs beyond its compiled code: the Android API level it
-// targets, the package name and human label, the requested runtime permissions, and the keystore
-// to sign with. The launcher activity is the compilation's entry point.
-case class ApkConfiguration
-  ( minApi:       Int,
-    targetApi:    Int,
-    packageName:  Text,
-    versionCode:  Int,
-    versionName:  Text,
-    label:        Text,
-    permissions:  List[Text],
-    keystore:     Text,
-    storePass:    Text,
-    alias:        Text,
-    keyPass:      Text )
