+++ lib/enigmatic/src/core/soundness_enigmatic_core.scala
@@ -41,8 +41,7 @@ export
       Des, Divulgence,
       Dsa, Ecb, Ecdsa, encrypt, Encryptor, Encryption,
       Hmac, hmac, InitializationVector, Iso10126, JavaStdlibCrypto, KeystoreError,
-      MlDsa, NoPadding, Ofb, Pem, PemError,
-      PemLabel, Password,
+      MlDsa, NoPadding, Ofb, Pem, Password,
       Permits, Pkcs7, PrivateKey, PublicKey, Rc2, Rsa, Signature, Signing,
       SignatureDigest, Symmetric, SymmetricKey, TripleDes, uncloak }
